Update TPACDiversityFundProposalComms.md
We should not promise a specific list of expenses for which the selection committee will make a grant.  We should at most provide some examples and encourage applicants to identify their needs.  There may be restrictions on the grants depending on how the funds are disbursed and any such restrictions will need to be considered by the selection committee.