Adjust order + content of identity characteristics
Due to the impact of race and class on equity, as well as the increasing awareness and resistance to indigenous groups and immigrants, members of the WCAG3 equity sub-group propose that these groups be added to the list of identities. Additionally, some members propose that the order be changed to reverse alpha-order. 

It is not clear if the characteristics were already in a deliberate order. Members of the equity sub-group are concerned that the lack of order may appear to imply a priority. By proposing a reverse alpha-order, we suggest this shows a conscious decision towards equity including by reversing the alphabetical order, since the traditional alpha-order often leads to inequity. If the reverse alpha-order or addition of characteristics have any concerns, please let me know, as I'm certainly open to discourse.

This PR adds the two characteristics, indigeneity and immigration status, and then orders the list in reverse alphabetical order.